Multiple NoiseFilters
° Adicionados diferentes tipos de NoiseFilters.

° Adição de ferramentas que auxiliarão e melhorarão a esfera.

° Refatoração de códigos.
