--- SpaceProject/Assets/Scripts/ColourSettings.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-[CreateAssetMenu]
-public class ColourSettings : ScriptableObject
-{
-    public Color planetColour;
-}
