+++ SpaceProject/Assets/Scripts/Colours/ColourSettings.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[CreateAssetMenu]
+public class ColourSettings : ScriptableObject
+{
+    public Color planetColour;
+}
