+++ SpaceProject/Assets/Scripts/Colours/ColourSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0054fd58b4f43c944bbd7854697d5430
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
