+++ SpaceProject/Assets/Scripts/ConditionalHideAtribute.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fc597f8d037802540bd1dc604ed3583c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
