--- SpaceProject/Assets/Scripts/Noise.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 448a103ab06c4e749b8766edb5618f67
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
