+++ SpaceProject/Assets/Scripts/Noise.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b5bd04914b0b987418dcb06f5b516d01
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
