+++ SpaceProject/Assets/Scripts/Noise/INoiseFilter.cs
@@ -1,0 +1,8 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface INoiseFilter 
+{
+    float Evaluate(Vector3 point);
+}
