+++ SpaceProject/Assets/Scripts/Noise/INoiseFilter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b918d52efd5fb0499a6a75758837694
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
