+++ SpaceProject/Assets/Scripts/Noise/NoiseFilterFactory.cs
@@ -1,0 +1,18 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public static class NoiseFilterFactory 
+{
+    public static INoiseFilter CreateNoiseFilter(NoiseSettings settings)
+    {
+        switch (settings.filterType)
+        {
+            case NoiseSettings.FilterType.Simple:
+                return new SimpleNoiseFilter(settings.simpleNoiseSettings);
+            case NoiseSettings.FilterType.Rigid:
+                return new RigidNoiseFilter(settings.rigidNoiseSettings);
+        }
+        return null;
+    }
+}
