+++ SpaceProject/Assets/Scripts/Noise/NoiseFilterFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ac3f49a4b1f0a841b7783f74ca15987
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
