+++ SpaceProject/Assets/Scripts/Noise/RigidNoiseFilter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 629f30c022569ad4b9e21fcc450888f3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
