+++ SpaceProject/Assets/Scripts/Noise/SimpleNoiseFilter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 87843d40dae90d9489ffca1a48a4c9a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
