--- SpaceProject/Assets/Scripts/NoiseSettings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7473b721766360747bab16845f60fddd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
