+++ SpaceProject/Assets/Scripts/Shape.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 56e552da8c3fbad4f9bd8470851fc44b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
