+++ SpaceProject/Assets/Scripts/Shape/ShapeSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 428010b64dd346c4495a8da6330f5838
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
