--- SpaceProject/Assets/Scripts/ShapeGenerator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e5c0163089240f348ad462d53b60a890
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
